requirements.txt
absl-py==0.7.0
astor==0.7.1
backcall==0.1.0
boto3==1.4.8
botocore==1.8.50
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.0
docutils==0.14
fire==0.1.3
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
jmespath==0.9.4
joblib==0.14.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
llvmlite==0.29.0
Markdown==3.0.1
matplotlib==3.0.3
mock==3.0.5
nltk==3.4.5
numba==0.41.0
numpy==1.16.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.10
protobuf==3.6.1
ptyprocess==0.6.0
py-rouge==1.1
Pygments==2.4.2
pyparsing==2.4.4
python-dateutil==2.8.0
pytz==2019.3
regex==2017.4.5
requests==2.22.0
s3transfer==0.1.13
sacremoses==0.0.35
scipy==1.1.0
sentencepiece==0.1.83
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
toposort==1.5
tqdm==4.31.1
traitlets==4.3.3
transformers==2.1.1
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.14.1
